inputstream file modularization
