+++ src/stx/CMakeLists.txt
@@ -164,14 +164,18 @@ add_library(stx-base STATIC
     inspect.cc
     InternMap.cc
     io/file.cc
+    io/FileInputStream.cc
     io/FileLock.cc
     io/fileutil.cc
     io/filerepository.cc
     io/inputstream.cc
     io/outputstream.cc
+    io/BufferInputStream.cc
     io/BufferedOutputStream.cc
+    io/MemoryInputStream.cc
     io/mmappedfile.cc
     io/pagemanager.cc
+    io/StringInputStream.cc
     ISO8601.cc
     CivilTime.cc
     Language.cc
