+++ src/stx/io/FileInputStream.h
@@ -1,0 +1,120 @@
+/**
+ * This file is part of the "libstx" project
+ *   Copyright (c) 2011-2014 Paul Asmuth, Google Inc.
+ *
+ * libstx is free software: you can redistribute it and/or modify it under
+ * the terms of the GNU General Public License v3.0. You should have received a
+ * copy of the GNU General Public License along with this program. If not, see
+ * <http://www.gnu.org/licenses/>.
+ */
+#ifndef _libstx_io_FileInputStream_h
+#define _libstx_io_FileInputStream_h
+#include <memory>
+#include <string>
+#include <stx/io/file.h>
+#include <stx/io/inputstream.h>
+
+namespace stx {
+
+class FileInputStream : public RewindableInputStream {
+public:
+  enum kByteOrderMark {
+    BOM_UNKNOWN,
+    BOM_UTF8,
+    BOM_UTF16
+  };
+
+  /**
+   * Open a new file input stream from the provided file path. Throws an
+   * exception if the file cannot be opened. The file will be automatically
+   * closed on destroy.
+   *
+   * @param file_path the path to the file
+   */
+  static std::unique_ptr<FileInputStream> openFile(
+      const std::string& file_path);
+
+  /**
+   * Create a new FileInputStream instance from the provided filedescriptor. If
+   * close on_destroy is true, the fd will be close()ed when the input stream
+   * is destroyed.
+   *
+   * @param fd a valid fd
+   * @param close_on_destroy close the fd on destroy?
+   */
+  static std::unique_ptr<FileInputStream> fromFileDescriptor(
+      int fd,
+      bool close_on_destroy = false);
+
+  /**
+   * Create a new FileInputStream instance from the provided File.
+   *
+   * @param file the opened file
+   */
+  static std::unique_ptr<FileInputStream> fromFile(File&& file);
+
+  /**
+   * Create a new FileInputStream instance from the provided filedescriptor. If
+   * close on_destroy is true, the fd will be close()ed when the input stream
+   * is destroyed.
+   *
+   * @param fd a valid fd
+   * @param close_on_destroy close the fd on destroy?
+   */
+  explicit FileInputStream(int fd, bool close_on_destroy = false);
+
+  /**
+   * Create a new FileInputStream instance from the provided File.
+   *
+   * @param file the opened file
+   */
+  explicit FileInputStream(File&& file);
+
+  /**
+   * Close the fd if close_on_destroy is true
+   */
+  ~FileInputStream();
+
+  /**
+   * Read the next byte from the file. Returns true if the next byte was read
+   * and false if the end of the stream was reached.
+   *
+   * @param target the target char pointer
+   */
+  bool readNextByte(char* target) override;
+
+  /**
+   * Skip the next N bytes in the stream. Returns the number of bytes skipped.
+   *
+   * @param n_bytes the number of bytes to skip
+   */
+  size_t skipNextBytes(size_t n_bytes) override;
+
+  /**
+   * Check if the end of this input stream was reached. Returns true if the
+   * end was reached, false otherwise
+   */
+  bool eof() override;
+
+  /**
+   * Rewind the input stream
+   */
+  void rewind() override;
+
+  /**
+   * Read the byte order mark of the file
+   */
+  kByteOrderMark readByteOrderMark();
+
+protected:
+  void readNextChunk();
+  char buf_[8192]; // FIXPAUL make configurable
+  size_t buf_len_;
+  size_t buf_pos_;
+  int fd_;
+  bool close_on_destroy_;
+};
+
+}
+
+#endif
