+++ src/stx/io/StringInputStream.h
@@ -1,0 +1,66 @@
+/**
+ * This file is part of the "libstx" project
+ *   Copyright (c) 2011-2014 Paul Asmuth, Google Inc.
+ *
+ * libstx is free software: you can redistribute it and/or modify it under
+ * the terms of the GNU General Public License v3.0. You should have received a
+ * copy of the GNU General Public License along with this program. If not, see
+ * <http://www.gnu.org/licenses/>.
+ */
+#ifndef _libstx_io_StringInputStream_h
+#define _libstx_io_StringInputStream_h
+#include <memory>
+#include <string>
+
+namespace stx {
+
+class StringInputStream : public RewindableInputStream {
+public:
+  /**
+   * Create a new InputStream from the provided string
+   *
+   * @param string the input string
+   */
+  static std::unique_ptr<StringInputStream> fromString(
+      const std::string& string);
+
+  /**
+   * Create a new InputStream from the provided string
+   *
+   * @param string the input string
+   */
+  StringInputStream(const std::string& string);
+
+  /**
+   * Read the next byte from the file. Returns true if the next byte was read
+   * and false if the end of the stream was reached.
+   *
+   * @param target the target char pointer
+   */
+  bool readNextByte(char* target) override;
+
+  /**
+   * Skip the next N bytes in the stream. Returns the number of bytes skipped.
+   *
+   * @param n_bytes the number of bytes to skip
+   */
+  size_t skipNextBytes(size_t n_bytes) override;
+
+  /**
+   * Check if the end of this input stream was reached. Returns true if the
+   * end was reached, false otherwise
+   */
+  bool eof() override;
+
+  /**
+   * Rewind the input stream
+   */
+  void rewind() override;
+
+protected:
+  std::string str_;
+  size_t cur_;
+};
+
+}
+#endif
