+++ src/stx/io/inputstream.cc
@@ -182,255 +182,4 @@ double InputStream::readDouble() {
   return IEEE754::fromBytes(val);
 }
 
-
-std::unique_ptr<FileInputStream> FileInputStream::openFile(
-    const std::string& file_path) {
-  auto fp = file_path.c_str();
-  int fd = open(fp, O_RDONLY);
-
-  if (fd < 1) {
-    RAISE_ERRNO(kIOError, "error opening file '%s'", fp);
-  }
-
-  std::unique_ptr<FileInputStream> stream(new FileInputStream(fd, true));
-  stream->readNextChunk();
-  stream->setFileName(file_path);
-  return stream;
-}
-
-std::unique_ptr<FileInputStream> FileInputStream::fromFileDescriptor(
-    int fd,
-    bool close_on_destroy /* = false */) {
-  std::unique_ptr<FileInputStream> stream(
-      new FileInputStream(fd, close_on_destroy));
-
-  stream->setFileName(StringUtil::format("<fd:$0>", fd));
-  return stream;
-}
-
-std::unique_ptr<FileInputStream> FileInputStream::fromFile(File&& file) {
-  auto fd = file.fd();
-
-  std::unique_ptr<FileInputStream> stream(new FileInputStream(std::move(file)));
-  stream->setFileName(StringUtil::format("<fd:$0>", fd));
-  return stream;
-}
-
-FileInputStream::FileInputStream(
-    int fd,
-    bool close_on_destroy /* = false */) :
-    fd_(fd),
-    close_on_destroy_(close_on_destroy),
-    buf_len_(0),
-    buf_pos_(0) {}
-
-FileInputStream::FileInputStream(
-    File&& file) :
-    FileInputStream(file.releaseFD(), true) {}
-
-FileInputStream::~FileInputStream() {
-  if (fd_ >= 0 && close_on_destroy_) {
-    close(fd_);
-  }
-}
-
-bool FileInputStream::readNextByte(char* target) {
-  if (buf_pos_ >= buf_len_) {
-    readNextChunk();
-  }
-
-  if (buf_pos_ < buf_len_) {
-    *target = buf_[buf_pos_++];
-    return true;
-  } else {
-    return false;
-  }
-}
-
-size_t FileInputStream::skipNextBytes(size_t nbytes) {
-  auto buf_remaining = buf_len_ - buf_pos_;
-
-  if (nbytes <= buf_remaining) {
-    buf_pos_ += nbytes;
-    return nbytes;
-  }
-
-  buf_pos_ = 0;
-  buf_len_ = 0;
-
-  if (lseek(fd_, nbytes - buf_remaining, SEEK_CUR) == -1) {
-    RAISE_ERRNO("seek() failed");
-  }
-
-  return nbytes;
-}
-
-bool FileInputStream::eof() {
-  if (buf_pos_ >= buf_len_) {
-    readNextChunk();
-  }
-
-  return buf_pos_ >= buf_len_;
-}
-
-// FIXPAUL move somwhere else...
-FileInputStream::kByteOrderMark FileInputStream::readByteOrderMark() {
-  if (buf_pos_ >= buf_len_) {
-    readNextChunk();
-  }
-
-  static char kByteOrderMarkUTF8[] = { '\xEF', '\xBB', '\xBF' };
-  if (buf_pos_ + sizeof(kByteOrderMarkUTF8) <= buf_len_ &&
-      strncmp(
-          buf_ + buf_pos_,
-          kByteOrderMarkUTF8,
-          sizeof(kByteOrderMarkUTF8)) == 0) {
-    buf_pos_ += sizeof(kByteOrderMarkUTF8);
-    return BOM_UTF8;
-  }
-
-  static char kByteOrderMarkUTF16[] = { '\xFF', '\xFE' };
-  if (buf_pos_ + sizeof(kByteOrderMarkUTF16) <= buf_len_ &&
-      strncmp(
-          buf_ + buf_pos_,
-          kByteOrderMarkUTF16,
-          sizeof(kByteOrderMarkUTF16)) == 0) {
-    buf_pos_ += sizeof(kByteOrderMarkUTF16);
-    return BOM_UTF16;
-  }
-
-  return BOM_UNKNOWN;
-}
-
-void FileInputStream::readNextChunk() {
-  int bytes_read = read(fd_, buf_, sizeof(buf_));
-
-  if (bytes_read < 0) {
-    RAISE_ERRNO(kIOError, "read(%s) failed", getFileName().c_str());
-  }
-
-  buf_pos_ = 0;
-  buf_len_ = bytes_read;
-}
-
-void FileInputStream::rewind() {
-  buf_pos_ = 0;
-  buf_len_ = 0;
-
-  if (lseek(fd_, 0, SEEK_SET) < 0) {
-    RAISE_ERRNO(kIOError, "lseek(%s) failed", getFileName().c_str());
-  }
-}
-
-std::unique_ptr<StringInputStream> StringInputStream::fromString(
-    const std::string& string) {
-  return std::unique_ptr<StringInputStream>(new StringInputStream(string));
-}
-
-StringInputStream::StringInputStream(
-    const std::string& string) :
-    str_(string),
-    cur_(0) {
-}
-
-bool StringInputStream::readNextByte(char* target) {
-  if (cur_ < str_.size()) {
-    *target = str_[cur_++];
-    return true;
-  } else {
-    return false;
-  }
-}
-
-size_t StringInputStream::skipNextBytes(size_t n_bytes) {
-  auto nskipped = n_bytes;
-  if (cur_ + n_bytes > str_.size()) {
-    nskipped = str_.size() - cur_;
-  }
-
-  cur_ += nskipped;
-  return nskipped;
-}
-
-bool StringInputStream::eof() {
-  return cur_ >= str_.size();
-}
-
-void StringInputStream::rewind() {
-  cur_ = 0;
-}
-
-std::unique_ptr<BufferInputStream> BufferInputStream::fromBuffer(
-    const Buffer* buf) {
-  return std::unique_ptr<BufferInputStream>(new BufferInputStream(buf));
-}
-
-BufferInputStream::BufferInputStream(
-    const Buffer* buf) :
-    buf_(buf),
-    cur_(0) {
-}
-
-bool BufferInputStream::readNextByte(char* target) {
-  if (cur_ < buf_->size()) {
-    *target = buf_->charAt(cur_++);
-    return true;
-  } else {
-    return false;
-  }
-}
-
-size_t BufferInputStream::skipNextBytes(size_t n_bytes) {
-  auto nskipped = n_bytes;
-  if (cur_ + n_bytes > buf_->size()) {
-    nskipped = buf_->size() - cur_;
-  }
-
-  cur_ += nskipped;
-  return nskipped;
-}
-
-bool BufferInputStream::eof() {
-  return cur_ >= buf_->size();
-}
-
-void BufferInputStream::rewind() {
-  cur_ = 0;
-}
-
-MemoryInputStream::MemoryInputStream(
-    const void* data,
-    size_t size) :
-    data_(data),
-    size_(size),
-    cur_(0) {}
-
-bool MemoryInputStream::readNextByte(char* target) {
-  if (cur_ < size_) {
-    *target = *(((char*) data_) + cur_++);
-    return true;
-  } else {
-    return false;
-  }
-}
-
-size_t MemoryInputStream::skipNextBytes(size_t n_bytes) {
-  auto nskipped = n_bytes;
-  if (cur_ + n_bytes > size_) {
-    nskipped = size_ - cur_;
-  }
-
-  cur_ += nskipped;
-  return nskipped;
-}
-
-bool MemoryInputStream::eof() {
-  return cur_ >= size_;
-}
-
-void MemoryInputStream::rewind() {
-  cur_ = 0;
-}
-
 } // namespace stx
-
