+++ src/stx/io/inputstream.h
@@ -18,7 +18,6 @@ namespace stx {
 
 class InputStream {
 public:
-
   /**
    * Get the stdin input stream
    */
@@ -159,251 +158,19 @@ private:
 
 class RewindableInputStream : public InputStream {
 public:
-
   /**
    * Rewind the input stream
    */
   virtual void rewind() = 0;
-
 };
 
-class FileInputStream : public RewindableInputStream {
-public:
-  enum kByteOrderMark {
-    BOM_UNKNOWN,
-    BOM_UTF8,
-    BOM_UTF16
-  };
+} // namespace stx
 
-  /**
-   * Open a new file input stream from the provided file path. Throws an
-   * exception if the file cannot be opened. The file will be automatically
-   * closed on destroy.
-   *
-   * @param file_path the path to the file
-   */
-  static std::unique_ptr<FileInputStream> openFile(
-      const std::string& file_path);
+// FIXME: this is just a quick hack to not break external deps
+// The app should include those headers directly instead.
+#include <stx/io/StringInputStream.h>
+#include <stx/io/BufferInputStream.h>
+#include <stx/io/MemoryInputStream.h>
+#include <stx/io/FileInputStream.h>
 
-  /**
-   * Create a new FileInputStream instance from the provided filedescriptor. If
-   * close on_destroy is true, the fd will be close()ed when the input stream
-   * is destroyed.
-   *
-   * @param fd a valid fd
-   * @param close_on_destroy close the fd on destroy?
-   */
-  static std::unique_ptr<FileInputStream> fromFileDescriptor(
-      int fd,
-      bool close_on_destroy = false);
-
-  /**
-   * Create a new FileInputStream instance from the provided File.
-   *
-   * @param file the opened file
-   */
-  static std::unique_ptr<FileInputStream> fromFile(File&& file);
-
-  /**
-   * Create a new FileInputStream instance from the provided filedescriptor. If
-   * close on_destroy is true, the fd will be close()ed when the input stream
-   * is destroyed.
-   *
-   * @param fd a valid fd
-   * @param close_on_destroy close the fd on destroy?
-   */
-  explicit FileInputStream(int fd, bool close_on_destroy = false);
-
-  /**
-   * Create a new FileInputStream instance from the provided File.
-   *
-   * @param file the opened file
-   */
-  explicit FileInputStream(File&& file);
-
-  /**
-   * Close the fd if close_on_destroy is true
-   */
-  ~FileInputStream();
-
-  /**
-   * Read the next byte from the file. Returns true if the next byte was read
-   * and false if the end of the stream was reached.
-   *
-   * @param target the target char pointer
-   */
-  bool readNextByte(char* target) override;
-
-  /**
-   * Skip the next N bytes in the stream. Returns the number of bytes skipped.
-   *
-   * @param n_bytes the number of bytes to skip
-   */
-  size_t skipNextBytes(size_t n_bytes) override;
-
-  /**
-   * Check if the end of this input stream was reached. Returns true if the
-   * end was reached, false otherwise
-   */
-  bool eof() override;
-
-  /**
-   * Rewind the input stream
-   */
-  void rewind() override;
-
-  /**
-   * Read the byte order mark of the file
-   */
-  kByteOrderMark readByteOrderMark();
-
-protected:
-  void readNextChunk();
-  char buf_[8192]; // FIXPAUL make configurable
-  size_t buf_len_;
-  size_t buf_pos_;
-  int fd_;
-  bool close_on_destroy_;
-};
-
-class StringInputStream : public RewindableInputStream {
-public:
-
-  /**
-   * Create a new InputStream from the provided string
-   *
-   * @param string the input string
-   */
-  static std::unique_ptr<StringInputStream> fromString(
-      const std::string& string);
-
-  /**
-   * Create a new InputStream from the provided string
-   *
-   * @param string the input string
-   */
-  StringInputStream(const std::string& string);
-
-  /**
-   * Read the next byte from the file. Returns true if the next byte was read
-   * and false if the end of the stream was reached.
-   *
-   * @param target the target char pointer
-   */
-  bool readNextByte(char* target) override;
-
-  /**
-   * Skip the next N bytes in the stream. Returns the number of bytes skipped.
-   *
-   * @param n_bytes the number of bytes to skip
-   */
-  size_t skipNextBytes(size_t n_bytes) override;
-
-  /**
-   * Check if the end of this input stream was reached. Returns true if the
-   * end was reached, false otherwise
-   */
-  bool eof() override;
-
-  /**
-   * Rewind the input stream
-   */
-  void rewind() override;
-
-protected:
-  std::string str_;
-  size_t cur_;
-};
-
-class BufferInputStream : public RewindableInputStream {
-public:
-
-  /**
-   * Create a new InputStream from the provided buffer
-   *
-   * @param string the input string
-   */
-  static std::unique_ptr<BufferInputStream> fromBuffer(const Buffer* buffer);
-
-  /**
-   * Create a new InputStream from the provided string
-   *
-   * @param string the input string
-   */
-  BufferInputStream(const Buffer* buffer);
-
-  /**
-   * Read the next byte from the file. Returns true if the next byte was read
-   * and false if the end of the stream was reached.
-   *
-   * @param target the target char pointer
-   */
-  bool readNextByte(char* target) override;
-
-  /**
-   * Skip the next N bytes in the stream. Returns the number of bytes skipped.
-   *
-   * @param n_bytes the number of bytes to skip
-   */
-  size_t skipNextBytes(size_t n_bytes) override;
-
-  /**
-   * Check if the end of this input stream was reached. Returns true if the
-   * end was reached, false otherwise
-   */
-  bool eof() override;
-
-  /**
-   * Rewind the input stream
-   */
-  void rewind() override;
-
-protected:
-  const Buffer* buf_;
-  size_t cur_;
-};
-
-class MemoryInputStream : public RewindableInputStream {
-public:
-
-  /**
-   * Create a new InputStream from the provided string
-   *
-   * @param string the input string
-   */
-  MemoryInputStream(const void* data, size_t size);
-
-  /**
-   * Read the next byte from the file. Returns true if the next byte was read
-   * and false if the end of the stream was reached.
-   *
-   * @param target the target char pointer
-   */
-  bool readNextByte(char* target) override;
-
-  /**
-   * Skip the next N bytes in the stream. Returns the number of bytes skipped.
-   *
-   * @param n_bytes the number of bytes to skip
-   */
-  size_t skipNextBytes(size_t n_bytes) override;
-
-  /**
-   * Check if the end of this input stream was reached. Returns true if the
-   * end was reached, false otherwise
-   */
-  bool eof() override;
-
-  /**
-   * Rewind the input stream
-   */
-  void rewind() override;
-
-protected:
-  const void* data_;
-  size_t size_;
-  size_t cur_;
-};
-
-}
 #endif
